+++ src/Controller/Policy/FileDownload.php
@@ -46,6 +46,7 @@ class FileDownload extends AbstractContainer
                 'folder' => 'plan_documents',
                 'idKey'  => 'plan_id',
             ];
+
             return PlanFile::find($info['id']);
         }
     }